refactor: split display_helpers.py (508 lines)
Extracted reference resolution into _display_references.py (143 lines).
Extracted preview helpers into _display_preview.py (114 lines).
display_helpers.py reduced from 509 to 251 lines.
All imports and __all__ preserved for backward compatibility.
Issue: #11
